agents/caption/requirements.txt
flask
gunicorn
torch
transformers
Pillow
